add a precompute pass
